This is the first WORKING version of the Throttle for XNET.
Currently, the throttle assumes 128 speed steps, and, it sets the throttle
speed to 0, with forward direction when control is first established (when the
first command is sent).

Additional changes made to include constants instead of having hard coded hex
values for all of the op codes.
Added file XNetConstants.java, which contains the constants.
Changed PowerManager to use the proper Lenz constants instead of the LocoNet
Constants it was using.
Changed the Turnout commands to use the constants instead of hard coded values.