Updated so that the add function now includes the Username.
Caught instances where if the systemname or turnout number was expected and if it didn't match with a specific prefix it would simply try to convert the value entered into a number, but would error if it contained a non-numerical character.

This now cross references to ensure that the username entered doesn't already exists as a system name and vice versa.

Tidied up some string handling and matching to convert it to uppercase, then only test against that.

Unable to properly test acela signalhead as it throws an error, but did so before I start work on changes.