A major update to the Throttle documentation, to be reviewed if possible. More doc for Jynstruments too. But also about the Media and Function labels panels in DecoderPro.
Draft of a howto towards building a Throttle oriented JMRI.

The general help index wasn't updated, can someone do it? (the ThrottleFrame.shtml in en/package/jmri/jmrit/throttle is a good index base)

I tried to merge some of the interesting content of the old throttle book  with this update, but I left a link to it, just in case.

Minor update to the code itself.

Everything can go in next major release.