Refactored Throttle Managers, so that only one throttle instance is created per loco address.
Dispatch, Release and Dispose are now handled by the Throttle Manager.  Exisiting methods of dispatching, releasing and dispatching within the throttle still exists but pass the requests back to the throttle manager.

Throttle Manager, now has an attach & remove Listener facility for code that only requires to be notified of changes to an address and no control.  A proxy method called getThrottleInfo is provided for code that does not require control, so that it can get the status of existing speed, direction function etc.