Refactored some things... streamlined the load process a bit.
VSDecoderManager now only reads the profile names
when the File->LoadVSDFile menu entry is called.

Loads the VSdecoder when a profile is selected
from the gui or when a roster entry is selected.

VSDecoders are unique per GUI now, but are still
held by the VSDecoderManager (and run on the 
manager's thread...)

Also gave each VSDecoder a unique ID based on
the SystemName convention...

Began the process of breaking the GUI away from
the guts.  Address changes are now passed via
a PropertyChangeEvent, though i don't like my
method of registering the events.

Some other housekeeping, removing dead code
and so on.