Convert as much as possible VSDecoder stuff to use LocoAddress instead of DccLocoAddress.
Have VSDecoderManager listen to all Reporters for loco entry/exit reports.
Forward messages to the right VSDecoder, to adjust the sound location to
the Reporter's position.

Right now this only works on Reporters added BEFORE the VSDecoderManager is loaded
(before the first VSDecoder is launched).

Add a dialog to the Debug menu to let users set the physical location of loaded
Reporters.  Location is stored in stringified form as a NamedBean Property.