Refactor the Active Train Frame to use Bundle.getMessage methods.
Make the start block combo box automatically select the first Occupied block in the list.

If the Active train frame has been called from the panel with a known roster entry and start block, populate the transit combo box with transits that only contain either the start block or have it as an entry block.