Add base support for IEEE 802.15.4 messages.  All this does at this point is allows sending and receiving of raw IEEE 802.15.4 messages from a serial port connection.
This configuration was tested using an XBee in API mode connected to the computer via a SparkFun Electronics XBee Explorer USB device.  Other devices supporting the IEEE 802.15.4 protocol should work as well, but have not been tested. 

The next step will be to add support for XBee IEEE 802.15.4 devices.

This code should be considered experimental.  There are likely bugs.