Add a JMRI schema catalog that can be added to an IDE to ensure that all JMRI XML schemas are referenced locally instead of at jmri.org.
In NetBeans, add this catalog at Tools -> DTDs and XML Schemas to allow NetBeans to use your local schema when validating an XML document in the editor. Note that this is global setting in NetBeans, so if you are working on mutiple JMRI branches, whichever branch you add the catalog from, that branch's schema is what is used. I have not testing this catalog in other IDEs.