Require a SystemConnectionMemo in the AbstractPortController constructor.
This is one of two steps towards using SystemConnectionMemos in all PortAdapters. The second step will be to require a non-null SystemConnectionMemo in the constructor. This is complicated by the number of different subclasses that expand upon the SystemConnectionMemo API.

After all systems use a non-null SystemConnectionMemo, the SystemConnectionMemo will get support for recognizing that it is "dirty" or carries unsaved configuration changes. Once the SystemConnectionMemo has that support, I can (finally I think) make the TabbedPreferences Save button behave correctly.