JmriPreferencesProvider provides managed Preferences. Shared (multi-computer) Preferences are stored in profile:profile/profile.properties Private (per-computer) Preferences are stored in profile:profile/<node-identity>/profile.properties
Profiles now used the shared Preferences to store identity and name and will automatically migrate from using profile:profile.properties (an xml file) to using the Shared Preferences. For now, Profiles will store the Profile identity and name in both the old and new formats.

This will impact users who changed the WebServer or JsonServer preferences using JMRI 4.1.1, but no provision for capturing those changes and migrating them to this structure is provided; it will be covered in the release notes, since this change is between test releases early in the test release cycle.