Commit of basic infrastructure for a JMRI Programmer implementation within the OpenLCB system.
This is a complete set of files, connects to tools like the 
Single CV Programmer, but does not yet function.  The next step
is to build out the operation code in OlcbProgrammer (and
OlcbProgrammerTest).

Note:  This code relies heavily in the openlcb library (see
http://openlcb.org/trunk/prototypes/java ) which is now orphaned.
As of 2015-09-03, version 0.6.4, the Memory Configuration Protocol in that
library is known to be broken. A patch has been submitted 
but there's no known maintainer for that library. Caveat Coder.