Fixes leaking debugmsg in case the user was pressing a different key than F8 at startup.
If the debugmsg is leaked, then the application hangs forever at startup. This was happening to me about 3 out of 4 startup attempts under linux. The typical case why another key event might be coming in is if the user operates the profiles dialog with the keyboard, all those events will show up here.