Add developerID attribute for LocoNet SV2 devices
Allows decoder files for LocoNet SV Format 2 devices to provide the device's SV2 developerID value as part of the <decoder><family><model> element.  The value is an integer, and may be used by JMRI SV2 tools when attempting to find a decoder file which corresponds to a device which has been reported by an SV2 Discovery reply message or an SV2 Identify reply message.