PreferencesProvider for Connections
This implementation is also a manager for ConnectionConfig (CC)
objects, although all CC objects are jointly registered using the old
(via the ConfigureManager and new (via the ConnectionConfigManager
(CCM)) mechanisms. CC object registered with the CCM are retained past
a save action in the TabbedPreferences, while the old registry is
cleared and not rebuilt (this has been confirmed to be true as far back
as 3.8, so I think this has always been true). The logic for really
adding and removing CCs from the CCM has not been wired into the
relevant PreferencesPanel yet.

All of the ConnectionConfigXml (CCX) objects have been refactored to
call a method of the CCX abstract object, so a single method contains
the register code, so future changes to registration are mostly in a
single source.