Update release process instructions
Added details from last release.
Clarified intent in a couple places.
More markdown formatting.
A bit of discussion about GitHub releases.

Note that this process now requires the release pumpkin to be able to write to GitHub JMRI/JMRI. This is because we don't have any other way to bring in a branch as a branch, instead of merging it onto master.