Replace deliberate NPE during error handling
Two routines (each file) had int return types, but were doing:

   return (Integer) null

which always throws an NPE during unboxing of the return value.  Apparently the author wrote "return null", couldn't get it to compile, didn't understand why, so added the cast.

Replaced with throw of an IllegalArgumentException, so at least we can find the problem easier when it happens.  Also added JavaDoc to the routine to document that.

Removed @SuppressWarnings("null") from these routines. That might case Jenkins to flag further items, but so be it.