More tightly link the Roster location to the UserFilesPath.
If the roster location is set to the UserFilesPath via any means,
future changes to the UserFilesPath cause the roster location to also
change.

This differs from the previous behavior that allowed users to use two
different methods to set the roster location to the UserFilesPath,
where one method would cause the roster location to change with the
UserFilePath (clicking the Reset button), and one method would not
allow that change to be followed (setting the path the same using the
file chooser).