Fix QualifiedVarTest
jmrit.symbolicprog.tabbedframe.QualifiedVarTest was failing due to missing  java/test/jmri/jmrit/decoderdefn/DecoderWithQualifier.xml file.  This came from another package, and a change to that package broke this test.

Did a bit of test-file refactoring: Moved a copy of the file here, and trimmed it up to improve this test.