Fix to NPE while configuring simulated connections
There seems to be several things going on that are causing various NPEs in the configuration process, see e.g. #232 and #236.

One is that, for some as-yet unknown reason, calls to AbstractSimulatorConnectionConfig.setInstance(..) that used to happen no longer are. That's the underlying cause of NPEs when AbstractSimulatorConnectionConfig.getConnectionName() is called. It seems to be hitting all simulator connections I've checked.

This bypasses the symptom by protecting against the NPE. But there's something more basic going on.