Some code cleanup.
d03430a changed the returned value of
```JmrixConfigPane.getCurrentObject()``` from ```Object``` to
```ConnectionConfig```. These changes reflect that.
Avoid double registering the first connection - it would show twice in
the Preferences window prior to removing the second registerPref call
in the FirstTimeStartUpWizard.

Note that these changes do not stop the preferences window from
displaying an empty pane for the first connection until after JMRI is
restarted. I suspect we need a repaint or similar call when the
TabbedPreferences window is opened to fix that.