Possibly fix NPEs saving connections
I think the problem with the NPEs saving connections was that when the
connection changed modes (from simulator to serial for example), the
old connection was not being removed from the ConnectionConfigManager
(CCM).

This has been fixed by implementing dispose() in the
AbstractConnectionConfig to remove the connection from the CCM and
having all subclasses call that implementation. Redundant overriding
implementations of dispose() have been removed.

This change also calls setInstance() whenever any ConnectionConfig is
registered. This sometimes throws errors, but they seem to be benign,
and may only indicate that code needs to be cleaned up.