Suppress warning
Annotate to suppress "unchecked" warning in use of InstanceManager.store()

Generally don't like to add these annotations, remove them when possible, but in this case the code uses an inferred type. I don't know how to code it so that the compiler understands that the class is handling it.  Hence the annotation and comment.