Annotate JFC casts
Jenkins test-warnings compile shows errors on test code using JFC Unit because of required casts.  Add annotations so they don't divert attention from other warnings.

Note: This is test code, so I felt it OK to annotate the whole (large) method instead of refactoring to put the JFC code in a smaller method.  But that's the better practice for the main code.