Set Python console encoding to UTF-8
Users can override this, but I can’t think of any value in doing so,
given that systems that don’t default to UTF-8 cause an error to be
logged, so this is not documented in python.properties.
Also allow users to override the python.cachedir. Again, this is not
documented in python.properties, since setting this incorrectly can
cause issues.