Set text black for cross-platform tests
This was failing on Appveyor, will see if this fixes it.

Also includes some clean up.

Tests known to fail during migration, like the rotated, animated GIF test, are not protected by a jmri.migrationtests parameter, but warn when it's not set to remind that they should eventually be fixed.  Set that parameter true to get them to run during migration work.