Update NW-1981 resource directories
The resources/icon/smallschematics/aspects/NW-1981 subdirectories had spaces in their names
- renamed them
- changed XML references to those

(This might break things for people e.g. using them standalone in panels, but shouldn't be an issue for users of the signal system)

Also, added the resource-tree index.php files so that the icons can be browsed.