Refactoring
This refactoring has no user-visible changes, but lays the ground work
for providing a single table of all startup actions that allows users
to reorder the actions within JMRI (so that, for example, a script
could be run both before and after a XML file is loaded). This also
removes a number of static lists.