Signal Controlled Warrants implemented. Warrants currently seems to be best suited for relatively long blocks, i.e. blocks with room enough for stepping the speed up and down and where the traffic is not so close that signals keeps changing frequently.
That leaves room for signal controlled warrants, where the speed of trains is changed more coarsely and where the train stops immediately if the next signal along the line changes to Stop and starts again when the next signal aspect allows that.

This commit is not intended to change any behaviour of the existing warrants, except for a little change in the way blocks are allocated: It is now secured that no deadlock occurs if two trains are running one behind the other, i.e that the train behind is not allocating a block in front of the first train.

The Edit warrant window has been enhanced with a choice between signal controlled warrants and normal warrants.

TODO:
1. A situation where the start block is not occupied by the train is not handled.
2. Even though the train will wait for signals, there can be situations where a turnout is not set correct and the train thereby leaves the designated route. That will make the train stop briefly, but not permanently as could be desired.