Clean // NOI18N comment
The "// NOI18N" comment means that a string on that line shouldn't be translated, e.g. its a translation key or something else that's _not_ presented to the user. Some IDEs need it just like that, which means that it shouldn't mix with _other_ comment text about that line.

This moves that magic comment to the end for one line in the file.