Refactor the power button code.
Cleans up a number of warnings in IDEs.
Removes a strange chain of dependencies on other widgets by directly
accessing the PowerManager when a user clicks the button.
Provides accessors to properties so this can be manipulated after
initialization (this would allow, for example, a future change to
show/hide labels in a toolbar).