Use RowSorter in UserPreferences instead of int.
Since the RowSorter provides a more powerful sorting mechanism that is
not directly related to the sort handlers in TableSorter, rewriting the
UserPreferences to use the RowSorter values instead of the TableSorter
values.
Make the RosterTable use a RowSorter instead of a TableSorter.
Make other components that store table sorting preferences use the
TableSorter.getSortOrder()/.getSortStatus() methods to translate
between the RowSorter and TableSorter mechanisms.