Do not save ProfileConfig.xml
ProfileConfig.xml is not compatible with the new startup action, which
will break earlier versions of JMRI, so we simply do not save it
anymore.
The new profile/profile.xml mechanism is compatible with schema changes
(as long as the namespace for each schema change differs), so there are
no worries there.