Remove technical details from online help.
The page WebServer.shtml documented security and technical details of
the server. Since many servlets are incapable of working within a
changing environment as outlined in the technical details, the security
content has been moved to index.shtml in the same directory as
WebServer.shtml and the technical content relegated to JavaDocs and
inline comments in the source code.
WebServer.shtml is now a direct redirection to index.shtml in the same
directory.