Do not log somewhat expected error.
The error in this case is the client stopped responding, but maintains
the connection, spamming the logs.
I haven’t figured out the cause and real fix in the JavaScript code,
but don’t need to spam anyone’s logs about this either.
(The debug statement in the onClose method does correctly report why
its closing as a result of this change, so we don’t need two logging
statements either).