Missing profile properties (#1514)
* Remove long-dead packages.

* Full name instead of Github user name.

* Fix error reading Profile preferences during Profile construction.

Attempts to read a Profile’s preferences during Profile construction
were flinging out on profiles created post 4.3.2 (as opposed to
profiles created pre 4.3.2). I did not completely understand or observe
this behavior until post 4.3.7. This was caused by passing an
incompletely constructed Profile object to a JmriPreferencesProvider in
an attempt to read the Profile’s preferences. The fix is to allow
JmriPreferencesProvider objects to be accessed with just the path to
the Profile instead of the complete Profile itself.
This also includes JavaDoc updates.

* Clarifying comment.

* Add tests for Profile