Do not call java with empty array on non-OS X systems.
Because I have barred the use of unset variables, an empty string (in
quotes) was being passed to the JVM on non-OS X systems when
DOCK_OPTIONS was empty.
Avoid this by not launching JMRI on non-OS X systems by performing a
check for OS X systems and calling java with different sets of
arguments depending on the OS (this is also how NetBeans launcher
handles this).