Remove apps.LocoTools
The LocoTools application was removed in JMRI 2.10 (June 2010), and
this package needs to be either removed or maintained to complete that.
The apps/LocoTools help content is heavily relied upon to describe
LocoNet tools, and makes no mention of the application, so its largely
left in place to avoid breaking links.