Silence warnings on missing tags
The logs on Travis-CI are *long* if warnings on missing tags are not silenced. They are in fact so long, they cannot be displayed in the Travis logs and there is a delay between failure (if it happens) and Travis-CI being able to provide the raw logs from AWS.
I will reintroduce this after significantly reducing the number of warnings in other branches.