Remove findbugs from ant call
On CI servers, need to use maven for findbugs since maven will ensure
findbugs is installed. This will cause the travis-ci build process to
move partly to maven so findbugs can be run within maven (the maven
findbugs plugin also has a handy switch for failing only on certain
priority bugs which is not present in the findbugs ant task, so that’s
a bonus).