Fix logic error in JRE detection on OS X
Unlike the error in 4.5.1, this error correctly detected the right version of java, but acted as if it did not. This change causes us not to unset JAVA_HOME after confirming that JAVA_HOME has the right version of java.
Also remove redundant check that JAVA_HOME/bin/java is executable.