Use standard JTable selection on editing behavior.
We inconsistently select the entire cell when starting an edit for
table cells (and this selection does not appear to function as
expected, at least on OS X). This change eliminates that inconsistency
and requires that where a non-standard behavior is desired that it be
implemented explicitly (it is probably best to implement the
non-standard behavior by providing a custom CellEditor for the column
needing the non-standard behavior instead of subclassing JTable).