MRC single bundle conversion
This time I think they will be in the right directory. Still have to go back and figure how you 'push' file deletions. There are going to be some orphan files after this. Also some of the translated properties have no translations in them (only english content) and would be cleaner to be deleted.