Do not error when we probably should not have.
Since the ShutDownManager API does not allow consumers to determine if
a task has been registered, do not penalize them for (re)registering a
task or for deregistering a task that is not registered. This allows
consumers to make a best effort to ensure tasks are (de)registered
without having to maintain separate tracking of task registration.