Fix incorrect OpSw31=c description
DCS240 OpSw31=c disables switch message metering, per the DCS240 user manual.  This has been confirmed with a physical DCS240.  

This is the opposite sense versus the setting for the DSC100 (per the DCS240 user manual and as confirmed with a physical DCS100).  It is likely that this file was initially copied from the DCS100options.properties file.