Remove JsonProgrammerServer
It’s broken, and not worth fixing since this is the last JSON server to
replace with a service factory.
This completes the modularization of the JSON services, such that with
a couple of historical artifacts in the JsonServlet, the only JSON code
any of the servers have is purely related to the core protocol.