Remove JARs published to maven central by upstream providers.
We don’t want to maintain JARs published to Maven Central by upstream
providers, since those files will be only downloaded once if in Maven
Central, but will be downloaded continuously if in our repository. The
validity of the download from Maven Central is hash summed (twice) and
signed using PGP, so its going to be legitimate.