Replace "Configuration Profiles" with "Projects"
I’ve never been happy with the name of the configuration profile
mechanism in documentation. Since it was referred to in at least three
different forms (“JMRI Profile”, “Configuration Profile”, and
“Profile”) and we have lots of different types of profiles (speed,
power, etc) under the hood, replace it with the terms “Project” or
“JMRI Project” (to distinguish it from another application’s project
type) in the user interfaces and documentation, but not in code (some
code used project when referring to Profile objects already).