Better handing of spaces in jmri.conf default options
Escaped spaces were being unescaped when passed to parse_args, and
wrapping default_options or JMRI_OPTIONS in quotes meant that options
were not being split at all. The solution is to use “read” to push
options into an array and then pass that array to parse_args. This
includes a short circuit of empty options.