No longer track JmrixConfigPanes for general use.
Since a JmrixConfigPane can only be placed in one AWT/Swing container
in its lifetime (if I understand how AWT/Swing works), do not attempt
to retain them for reuse. This means a number of now-deprecated methods
in JmrixConfigPane will show that there are no connections configured.
As a result:
- Update AppConfigBase to check for duplicates using ConnectionConfigs
instead of JmrixConfigPanes (a JmrixConfigPane is the GUI for a
ConnectionConfig)
- Update ConnectionsPreferencesPanel to create its set of
JmrixConfigPanes from ConnectionConfigManager’s collection of
connections and to listen to changes in the ConnectionConfigManager.
- Update the new project Wizard to create a JmrixConfigPane for its own
use and to save preferences directly.