AbstractMonitor file selection improvements (#2620)
Creates a modal dialog box if FileOutputStream() throws a FileNotFoundException when
"Start logging" is activated.  Previous code simply left an error and exception info
in the console log without informing the user of the failure in opening the file for
writing.

Ensures that the default directory for MonitorLog.txt files is the profile directory.  This specifically applies:
- when user doesn't use the "Choose file name" button
- when user opens "Choose file name" but uses the "cancel" button.

Improves JavaDocs within AbstractMonPane.java

	modified:   java/src/jmri/jmrix/AbstractMonPane.java
	modified:   java/src/jmri/jmrix/JmrixBundle.properties