Fix pause being added to startup even when cancelled.
When a user adds a pause to the startup sequence, the pause gets added
even if the user cancels adding it. This was caused by the pause always
returning a non-null, non-empty name for the action.

The fix is to add an ```isValid()``` method to StartupModels and test
that for inclusion in the startup sequence and to make sure that the
default state of a pause action is invalid, so that when a user cancels
adding the pause, an invalid pause is passed to the preferences panel,
which drops it (ideally the manager should drop it, but we want to
allow invalid actions while loading…)