Begin making WarrantPreferences a preferences store
The Warrant objects are all tightly coupled, expecting to be able to
set static values in each other. Begin switching to a model where the
preferences are listened to for changes by interested objects instead
of a model where the preferences reaches out to adjust those objects it
knows needs adjustment.