Only initialize fields if needed.
A side note: Java servlet initialization is very lazy; a servlet will
only be initialized when first needed (even though it remains in memory
until used). Since these servlets support multiple paths, only perform
a complete initialization for paths that require it. This may be
overkill in these cases (I don’t think the ObjectMapper requires much
resources), but they are useful examples of initialization done
correctly (the call to super.init() was removed based on the
GenericServlet documentation—init() is an empty convenience method in
GenericServlet that exists only so that init methods with parameters
can call it after setting up the servlet’s context.