Include roster entry attributes in JSON messages
Also ensure last operated and last updated values are stored as ISO
8601 dates since JavaScript requires that dates be stored that way
(this allows the web services to consume them). This has the side
effect of allowing last modified dates to be sorted in the roster table
if the stored last modified date is parsable as a Date (it is otherwise
not displayed in the table, but is displayed in the roster entry
editors).