Allow the StartupManager to not perform actions when loading preferences
The StartupManager now does a two step process for loading its
preferences. It loads all preferences into memory (creating the list of
actions available in the preferences window) and then, if there were
not errors blocking it, performs those actions.

To make this work, the triggers for startup actions were moved from the
XmlAdapters that load the actions to the StartupModels themselves. Some
actions still need to throw nicer error messages, but this generally
works now.