Allow the web server life cycle to be monitored externally
By extending the Jetty AbstractLifeCycle, other JMRI components can
listen to the lifecycle of the JMRI web server. This would allow, for
example, another JMRI service to only take a potentially expensive
action if the user starts the web server, without ever having to make
the web server aware of that other service.