Updated ARNOLD 81210-81 decoder definition xml
In the former Arnold.xml file definition I created in July 2010 (years ago), the consist CV (not documented by Arnold) was missing as it was not documented in the Arnold manual. However when resetting the decoder it was set to 255 (Consist address 127, reverse).
I opened issue #2921 including a zip with the updated xml, but I think it was not the right place to do it. Thus this pull request.
Thanks to validate and include in a future release.
Alain