Significant cleanup / bug-fix rewrite of the ConfigBaseStationFrame.
Extracted the common code for the three TableModel inner classes
to a new shared DCCppTableModel class.

Stripped out a bunch of junk code.

Fixed the strange behavior with the table's row  numbering and
the Model's row numbering getting out of sync.

Fixed the Delete button behavior.

Added the option to Write to EEPROM when  you click the "Save" button
not just the Close button.