Warn instead of error if class not found
The ```setPreferenceState(className, prefItem, state)``` method caused
an error and stack trace to be logged if className was not a class. The
Javadocs indicate that className need not be resolvable, only that
className + prefItem be unique. This change warns if the className does
not resolve to a class, since this allows a potential warning that a
plugin may not be present.