Fail tests with unexpected error messages within the test
Assert an unexpected error message within the test logging the message
(or the first test after it to assert this). Retain the separate test
as a potential failsafe (if run after causing test). This is needed to
ensure that error messages are caught when running individual tests and
so that tests can be run in random order on CI servers.