Reliably reset operations tests to default state
Operations setup was a library of static methods accessing static
variables; refactoring it into a library of static methods accessing a
default instance allows us to blow away the default instance when tests
are complete to reset the operations setup.

This solves problems when an unidentifiable test leaves the operations
setup in a non-default state.