Do not display connection errors as XML panel errors
Do not use the default, panel XML oriented, XML error handled, but use
a custom error handler that captures the error and retains it until
startup errors are displayed. The captured error is shown indented from
the “Unable to create connection X” error.
Also memos containing serial port not found and busy errors are
localizable (the mechanism needs to be different though)