Adds force notify for turnout object state changes. This will ensure that events get generated to the bus even if the last known state matches the new event; in essence this creates duplicate events on the bus.
This feature is necessary in case the JMRI state has gone out of sync from the bus state and the user wants to reinitialize all turnouts/etc with logix.
It can also be helpful in case multiple turnout objects are used to create a single multi-aspect signal head, with each turnout and each state representing a unique aspect, but changes possibly going between turnouts.