Bundle name null issue
JavaDoc warnings about name being null if no properties file, but name
is declared to be international, so can't be null, and parent also says
it can't be null. So all Bundle.java point to a Bundle.properties, even
if it is only a file with a header comment.
This clears the warning and sets the place for later strings that likely
hide in the code that should be translated.