Retain list of JmriJFrames within InstanceManager
This ensures that when running tests in a random order within the same
JVM, that resetting the InstanceManager also means that the list of all
frames gets reset as well. Not resetting the list of frames was causing
tests that followed a test that created a TabbedPreferencesFrame were
consistently requiring the TabbedPreferences default instance be
recreated.