Call AboutDialog realisticly
The AboutDialog is created without a parent frame in JMRI applications,
so do so in the tests as well. This also allows tests to watch the
entire display for the dialog instead of just the frame. Also add an
explicit failure if dialog constructor returns null or if dialog cannot
be closed.