Ignore the ProcessSerialNumber argument
The ```open``` command on macOS adds a ProcessSerialNumber argument (```-psn_#_########```) as the last argument when running a JMRI bundled application (this may also occur if the application is launched directly from a DMG while double clicking it in the Finder). This ignores that argument so JMRI app does not attempt to interpret it as an alternate configuration name.