Prevent potential duplicate user names
The turnout, sensor and light tables can have multiple tabs based on
the number of connections.  Except for the All tab, these tabs have a
manager with a subset of the objects.  The Rename process was using the
tab specific manager instead of the global manager to check for
duplicate user names.