DCC4PC Message Handling
Number of improvements to improve the robustness the DCC-4-PC code.

Reader board discovery is now in its own class so that the message/reply sequence no longer interferes with the polling messages.

Solicited replies now contain the original message.  As the reply contains no information as to its source this removes the requirement for the message sender to try and keep track of what messages it has sent and had or missed replies for.  This information can now be retrieved from the reply and, removes any issues of sender getting out of sync with board poll status requests and the replies.