LayoutEditor: Changed meta key to right mouse button on Linux
Changed the method isMetaDown() to also check for Linux and if so, returns SwingUtilities.isRightMouseButton(event).

This is because the meta key isn't the right mouse button anymore on Java 9 on Linux.

Layout editor: Meta key does not work on Java 9 on Windows 10.
The same problem is on Linux and this PR fixes that.