BugFix for TreeModel
The problem is that getNode keyed off of the component… but for controllers that's always null… so one node for all controllers... not good. If you key off of controller instead the that's the same for all components... again, not good. The fix is to key off the name; which seems to be unique between all controllers and components… (except for "UNKNOWN [UNKNOWN]"'s)… but I'm thinking that those should be ignored anyway...