Add API for generating list of objects with expected state
JMRI currently contains a number of situations where it generates, within a method, two lists, one of which is returned and one of which is retained for immediate retrival from another method. This structure is fundanmentally broken in that a method that be design does not change state does change state.

This API is designed to capture an object (normally a NamedBean or NamedBean wrapper) and its expected state (using the NamedBean state as the model for expected state).