Remove bad tests
Remove tests that instances of classes that are documented as being uninstantiatable can be created.
Remove tests that mask incorrectly written classes, which would show as being tested if correctly written, despite having no explicit tests written against them (the lack of test coverage for the class indicates it is incorrectly written).