Update index.shtml
Added missing item to list to correspond to index in the sidebar (on left side of page when viewed on web site).  [[Query:  what file is that sidebar in?  I have some ideas to split and add pages to the whole section on scripting to make it more usable, but I want to keep that sidebar index in sync with the pages I would split and/or create.  Thanks.]] 

Jerry