Change SignalSystemTest due to change in DefaultSignalMastLogicManager
DefaultSignalMastLogicManager now inherits AbstractManager instead of keeping its own list of beans. Before, DefaultSignalMastLogicManager stored the beans in an unsorted list, but now when it inherits AbstractManager, the beans are stored in a sorted set.

In this test, 20 different error messages are reported. But it's only the first error message that is checked. And since the beans are now sorted, the error messages come in a different order.