Update LAISDCC definition
A newer serie of decoders, named KungFu (8700xx instead of 8600xx for the PanGu serie) are available, with same set of CVs, but with a higher Max Amp power (up to 2 Amp, instead of 1 Amp). their version ID is higher (6 in the one I received). Thus in this update I updated highVersionID to 06, and changed type of decoders to 8x00nn instead of 8600nn