Add preferences to allow loopback and linklocal addresses for ZeroConf services.
This is *not* exposed to users, but can be used where testing is required. This is added because of discovery that WiThrottle server is really screwing up requesting JmDNS advertisements in IPv6 network, but only real way to test on some networks is to change code to allow link local addresses to be used.