Fixes JsonRosterSocketServiceTest having random failures on CI in the testListen method.
The problem was that during the initialization some property update notifications are enqueued
to the AWT thread. If the AWT thread falls behind due to random kernel scheudling,
these property update notifications turned into JSON messages which caused havoc
against the expectations.

Due to a too strict wording of the waitFor condition (size == 1 instead of size >=1),
the error was "message not sent" when in fact there were too many messages sent.